# fix(perf): filtering process in application-list api (#12985)