--- a/util/argo/argo.go
+++ b/util/argo/argo.go
@@ -63,6 +63,26 @@ func FilterByProjects(apps []argoappv1.Application, projects []string) []argoapp
 
 }
 
+// FilterByProjectsP returns application pointers which belongs to the specified project
+func FilterByProjectsP(apps []*argoappv1.Application, projects []string) []*argoappv1.Application {
+	if len(projects) == 0 {
+		return apps
+	}
+	projectsMap := make(map[string]bool)
+	for i := range projects {
+		projectsMap[projects[i]] = true
+	}
+	items := make([]*argoappv1.Application, 0)
+	for i := 0; i < len(apps); i++ {
+		a := apps[i]
+		if _, ok := projectsMap[a.Spec.GetProject()]; ok {
+			items = append(items, a)
+		}
+	}
+	return items
+
+}
+
 // FilterAppSetsByProjects returns applications which belongs to the specified project
 func FilterAppSetsByProjects(appsets []argoappv1.ApplicationSet, projects []string) []argoappv1.ApplicationSet {
 	if len(projects) == 0 {
@@ -96,6 +116,20 @@ func FilterByRepo(apps []argoappv1.Application, repo string) []argoappv1.Applica
 	return items
 }
 
+// FilterByRepoP returns application pointers
+func FilterByRepoP(apps []*argoappv1.Application, repo string) []*argoappv1.Application {
+	if repo == "" {
+		return apps
+	}
+	items := make([]*argoappv1.Application, 0)
+	for i := 0; i < len(apps); i++ {
+		if apps[i].Spec.GetSource().RepoURL == repo {
+			items = append(items, apps[i])
+		}
+	}
+	return items
+}
+
 // FilterByCluster returns an application
 func FilterByCluster(apps []argoappv1.Application, cluster string) []argoappv1.Application {
 	if cluster == "" {
@@ -116,6 +150,22 @@ func FilterByName(apps []argoappv1.Application, name string) ([]argoappv1.Applic
 		return apps, nil
 	}
 	items := make([]argoappv1.Application, 0)
+	for i := 0; i < len(apps); i++ {
+		if apps[i].Name == name {
+			items = append(items, apps[i])
+			return items, nil
+		}
+	}
+	return items, status.Errorf(codes.NotFound, "application '%s' not found", name)
+}
+
+// FilterByNameP returns pointer applications
+// This function is for the changes in #12985.
+func FilterByNameP(apps []*argoappv1.Application, name string) ([]*argoappv1.Application, error) {
+	if name == "" {
+		return apps, nil
+	}
+	items := make([]*argoappv1.Application, 0)
 	for i := 0; i < len(apps); i++ {
 		if apps[i].Name == name {
 			items = append(items, apps[i])