--- a/util/argo/argo.go
+++ b/util/argo/argo.go
@@ -161,18 +161,18 @@ func FilterByName(apps []argoappv1.Application, name string) ([]argoappv1.Applic
 
 // FilterByNameP returns pointer applications
 // This function is for the changes in #12985.
-func FilterByNameP(apps []*argoappv1.Application, name string) ([]*argoappv1.Application, error) {
+func FilterByNameP(apps []*argoappv1.Application, name string) []*argoappv1.Application {
 	if name == "" {
-		return apps, nil
+		return apps
 	}
 	items := make([]*argoappv1.Application, 0)
 	for i := 0; i < len(apps); i++ {
 		if apps[i].Name == name {
 			items = append(items, apps[i])
-			return items, nil
+			return items
 		}
 	}
-	return items, status.Errorf(codes.NotFound, "application '%s' not found", name)
+	return items
 }
 
 // RefreshApp updates the refresh annotation of an application to coerce the controller to process it