@blakepettersson 2023-03-24
How can we assure that 

1) This is faster than what we have now
2) That this will remain fast if 1) is indeed true

i.e. are there any benchmarks or can those somehow be provided?