@blakepettersson 2023-03-24
> Or, do you have a better and best way to conduct benchmarking?

Sounds good to me 😄 

> And does this need to be reproducible?

Ideally I'd like to be able to run a `BenchmarkApplicationList` method, but IMO not a showstopper if that can't be provided.