@tken2039 2023-03-27
> BenchmarkApplicationList

Thank you for your advice.
I am currently looking into implementing this one. Please wait a little.