@tken2039 2023-03-27
Sorry for the delay.
I have run the benchmark in the following format.

As a result, I have found no significant processing time issues with this PR.

| Pattern No. | Number of Applications | Specify Application Name in Query | Specify Projects in Query | Specify Repo in Query | Note |
|-|-|-|-|-|-|
| 1 | 10 | x | x | x | Function name: BenchmarkListFewApps |
| 2 | 500 | x | x | x | Function name: BenchmarkListSomeApps |
| 3 | 10000 | x | x | x | Function name: BenchmarkListMuchApps |
| 4 | 10000 | o | x | x | Function name: BenchmarkListMuchAppsWithName |
| 5 | 10000 | x | o | x | Function name: BenchmarkListMuchAppsWithProjects |
| 6 | 10000 | x | x | o | Function name: BenchmarkListMuchAppsWithRepo |

## Benchmark

To verify the processing speed of the code without this modification, the following branch is created.
https://github.com/tken2039/argo-cd/pull/1

<details>
<summary>Run benchmark(Click to open)</summary>

- Not adding this PR changes

  ```sh
  ~/work/tmp/argo-cd/server/application tmp/add-benchmark-for-application-list
  ❯ go test -benchmem -run='^$' -bench=. > ~/work/tmp/before-12999.out                  

  ~/work/tmp/argo-cd/server/application tmp/add-benchmark-for-application-list 1m 11s
  ❯
  ```

- Adding this PR changes

  ```sh
  ~/work/tmp/argo-cd/server/application feature/replase-filtering-by-query ⇣⇡
  ❯ go test -benchmem -run='^$' -bench=. > ~/work/tmp/after-12999.out

  ~/work/tmp/argo-cd/server/application feature/replase-filtering-by-query ⇣⇡ 41s
  ❯ 
  ```

</details>

<details>
<summary>Results And Comparison(Click to open)</summary>

Unnecessary logs are deleted from each file.

- Not adding this PR changes

  ```sh
  ❯ cat before-12999.out  
  goos: darwin
  goarch: arm64
  pkg: github.com/argoproj/argo-cd/v2/server/application
  BenchmarkListMuchApps-10                              31          33733034 ns/op        68801303 B/op     490908 allocs/op
  BenchmarkListSomeApps-10                            2386            501249 ns/op         1233498 B/op       5775 allocs/op
  BenchmarkListFewApps-10                           104238             11004 ns/op           36862 B/op        128 allocs/op
  BenchmarkListMuchAppsWithName-10                      36          28954932 ns/op        66060855 B/op     437981 allocs/op
  BenchmarkListMuchAppsWithProjects-10                  38          28841561 ns/op        65166212 B/op     420739 allocs/op
  BenchmarkListMuchAppsWithRepo-10                      34          29688271 ns/op        67050388 B/op     457279 allocs/op
  PASS
  ok      github.com/argoproj/argo-cd/v2/server/application       65.699s

  ~/work/tmp
  ❯ 
  ```

- Adding this PR changes

  ```sh
  ❯ cat after-12999.out                             
  goos: darwin
  goarch: arm64
  pkg: github.com/argoproj/argo-cd/v2/server/application
  BenchmarkListMuchApps-10                              33          32714341 ns/op        67610686 B/op     467803 allocs/op
  BenchmarkListSomeApps-10                            2348            490419 ns/op         1233709 B/op       5779 allocs/op
  BenchmarkListFewApps-10                           105544             11320 ns/op           36861 B/op        128 allocs/op
  BenchmarkListMuchAppsWithName-10                    6600            213013 ns/op          564007 B/op        429 allocs/op
  BenchmarkListMuchAppsWithProjects-10                3606            364228 ns/op          601513 B/op        769 allocs/op
  BenchmarkListMuchAppsWithRepo-10                    3700            310253 ns/op          596047 B/op        736 allocs/op
  PASS
  ok      github.com/argoproj/argo-cd/v2/server/application       35.896s

  ~/work/tmp
  ❯ 
  ```


### Comparison of Benchmark Results

```
~/work/tmp
❯ benchstat before-12999.out after-12999.out
goos: darwin
goarch: arm64
pkg: github.com/argoproj/argo-cd/v2/server/application
                            │ before-12999.out │            after-12999.out            │
                            │      sec/op      │    sec/op     vs base                 │
ListMuchApps-10                   33.73m ± ∞ ¹   32.71m ± ∞ ¹        ~ (p=1.000 n=1) ²
ListSomeApps-10                   501.2µ ± ∞ ¹   490.4µ ± ∞ ¹        ~ (p=1.000 n=1) ²
ListFewApps-10                    11.00µ ± ∞ ¹   11.32µ ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithName-10         28954.9µ ± ∞ ¹   213.0µ ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithProjects-10     28841.6µ ± ∞ ¹   364.2µ ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithRepo-10         29688.3µ ± ∞ ¹   310.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                           4.080m         404.4µ        -90.09%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                            │ before-12999.out │            after-12999.out             │
                            │       B/op       │     B/op       vs base                 │
ListMuchApps-10                  65.61Mi ± ∞ ¹   64.48Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
ListSomeApps-10                  1.176Mi ± ∞ ¹   1.177Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
ListFewApps-10                   36.00Ki ± ∞ ¹   36.00Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithName-10        64512.6Ki ± ∞ ¹   550.8Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithProjects-10    63638.9Ki ± ∞ ¹   587.4Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithRepo-10        65478.9Ki ± ∞ ¹   582.1Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                          9.376Mi         902.2Ki        -90.60%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                            │ before-12999.out │            after-12999.out            │
                            │    allocs/op     │  allocs/op    vs base                 │
ListMuchApps-10                   490.9k ± ∞ ¹   467.8k ± ∞ ¹        ~ (p=1.000 n=1) ²
ListSomeApps-10                   5.775k ± ∞ ¹   5.779k ± ∞ ¹        ~ (p=1.000 n=1) ²
ListFewApps-10                     128.0 ± ∞ ¹    128.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
ListMuchAppsWithName-10         437981.0 ± ∞ ¹    429.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithProjects-10     420739.0 ± ∞ ¹    769.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
ListMuchAppsWithRepo-10         457279.0 ± ∞ ¹    736.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                           55.92k         2.093k        -96.26%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal

~/work/tmp
❯ 
```

</details>