@tken2039 2023-03-28
Added error checking for lint.

de7186b